[feat] 마이픽 -> (여행) 상세 정보 화면 연동, 마이픽 아이템 클릭 이벤트 구현

+++ feature/mypage/src/main/kotlin/com/tripmate/android/feature/mypage/viewmodel/MyPageUiEvent.kt
@@ -11,7 +11,7 @@ sealed interface MyPageUiEvent {
     data object NavigateToMyPick : MyPageUiEvent
     data object NavigateToLogin : MyPageUiEvent
     data object NavigateToWithdraw : MyPageUiEvent
-    data class NavigateToTripDetail(val spotId: Long) : MyPageUiEvent
+    data class NavigateToTripDetail(val spotId: Int) : MyPageUiEvent
     data object Withdraw : MyPageUiEvent
     data object NavigateToMain : MyPageUiEvent
     data object NavigateToPrivacyPolicy : MyPageUiEvent
